# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachRequest": {
        "properties": {
          "page_id": {
            "title": "Page Id",
            "type": "string"
          }
        },
        "required": [
          "page_id"
        ],
        "title": "AttachRequest",
        "type": "object"
      },
...
      "Brief": {
        "properties": {
          "audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Audience"
          },
          "business": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Business"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Goal"
          },
          "tone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tone"
          }
        },
        "title": "Brief",
        "type": "object"
      },
      "BuildDetail": {
        "properties": {
          "brief": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Brief"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/SavedMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "pageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pageid"
          },
          "pageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pageurl"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "turns": {
            "title": "Turns",
            "type": "integer"
          },
          "updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated"
          }
        },
        "required": [
          "id",
          "title",
          "turns",
          "messages"
        ],
        "title": "BuildDetail",
        "type": "object"
      },
      "BuildList": {
        "properties": {
          "builds": {
            "items": {
              "$ref": "#/components/schemas/BuildSummary"
            },
            "title": "Builds",
            "type": "array"
          }
        },
        "required": [
          "builds"
        ],
        "title": "BuildList",
        "type": "object"
      },
      "BuildPatch": {
        "properties": {
          "brief": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Brief"
              },
              {
                "type": "null"
              }
            ]
          },
          "messages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SavedMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messages"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "BuildPatch",
        "type": "object"
      },
      "BuildSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pageid"
          },
          "pageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pageurl"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "turns": {
            "title": "Turns",
            "type": "integer"
          },
          "updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated"
          }
        },
        "required": [
          "id",
          "title",
          "turns"
        ],
        "title": "BuildSummary",
        "type": "object"
      },
...
      "Deleted": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "string"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "Deleted",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Current"
          },
          "expected": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Expected"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "readiness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Readiness"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "recognizedCommand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recognizedcommand"
          },
          "validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GoogleLogin": {
        "properties": {
          "idToken": {
            "title": "Idtoken",
            "type": "string"
          }
        },
        "required": [
          "idToken"
        ],
        "title": "GoogleLogin",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginResponse": {
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "OpenBuild": {
        "properties": {
          "build": {
            "$ref": "#/components/schemas/BuildDetail"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageView"
              },
              {
                "type": "null"
              }
            ]
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionEntry"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "build",
          "versions"
        ],
        "title": "OpenBuild",
        "type": "object"
      },
...
      "PageView": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Name"
          },
          "published": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublishedRef"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "readiness": {
            "$ref": "#/components/schemas/Readiness"
          },
          "sections": {
            "items": {
              "discriminator": {
                "mapping": {
                  "about": "#/components/schemas/AboutSection",
                  "benefits": "#/components/schemas/BenefitsSection",
                  "contact": "#/components/schemas/ContactSection",
                  "cta": "#/components/schemas/CtaSection",
                  "faq": "#/components/schemas/FAQSection",
                  "footer": "#/components/schemas/FooterSection",
                  "gallery": "#/components/schemas/GallerySection",
                  "header": "#/components/schemas/HeaderSection",
                  "hero": "#/components/schemas/HeroSection",
                  "pricing": "#/components/schemas/PricingSection",
                  "process": "#/components/schemas/ProcessSection",
                  "promotion": "#/components/schemas/PromotionSection",
                  "services": "#/components/schemas/ServicesSection",
                  "stats": "#/components/schemas/StatsSection",
                  "team": "#/components/schemas/TeamSection",
                  "testimonials": "#/components/schemas/TestimonialsSection"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/HeaderSection"
                },
                {
                  "$ref": "#/components/schemas/HeroSection"
                },
                {
                  "$ref": "#/components/schemas/AboutSection"
                },
                {
                  "$ref": "#/components/schemas/BenefitsSection"
                },
                {
                  "$ref": "#/components/schemas/ServicesSection"
                },
                {
                  "$ref": "#/components/schemas/ProcessSection"
                },
                {
                  "$ref": "#/components/schemas/StatsSection"
                },
                {
                  "$ref": "#/components/schemas/GallerySection"
                },
                {
                  "$ref": "#/components/schemas/TestimonialsSection"
                },
                {
                  "$ref": "#/components/schemas/PricingSection"
                },
                {
                  "$ref": "#/components/schemas/TeamSection"
                },
                {
                  "$ref": "#/components/schemas/PromotionSection"
                },
                {
                  "$ref": "#/components/schemas/FAQSection"
                },
                {
                  "$ref": "#/components/schemas/ContactSection"
                },
                {
                  "$ref": "#/components/schemas/CtaSection"
                },
                {
                  "$ref": "#/components/schemas/FooterSection"
                }
              ]
            },
            "title": "Sections",
            "type": "array"
          },
          "theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Theme"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "url",
          "sections",
          "readiness"
        ],
        "title": "PageView",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "SavedMessage": {
        "properties": {
          "clarification": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarification"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "tone": {
            "anyOf": [
              {
                "const": "error",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone"
          }
        },
        "required": [
          "role",
          "text"
        ],
        "title": "SavedMessage",
        "type": "object"
      },
...
      "TokenPair": {
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ],
        "title": "TokenPair",
        "type": "object"
      },
...
      "User": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "picture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Picture"
          }
        },
        "required": [
          "id"
        ],
        "title": "User",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationStatus": {
        "properties": {
          "gates": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Gates"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "ValidationStatus",
        "type": "object"
      },
      "VersionEntry": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "published": {
            "title": "Published",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "when": {
            "title": "When",
            "type": "string"
          }
        },
        "required": [
          "id",
          "version",
          "label",
          "when",
          "published"
        ],
        "title": "VersionEntry",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Magic Landing Pages",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/google": {
      "post": {
        "operationId": "google_login_auth_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me_auth_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refresh_tokens_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/builds": {
      "get": {
        "operationId": "list_builds_builds_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Builds",
        "tags": [
          "builds"
        ]
      },
      "post": {
        "operationId": "create_build_builds_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Build",
        "tags": [
          "builds"
        ]
      }
    },
    "/builds/{build_id}": {
      "delete": {
        "operationId": "delete_build_builds__build_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Build",
        "tags": [
          "builds"
        ]
      },
      "get": {
        "operationId": "get_build_builds__build_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenBuild"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Build",
        "tags": [
          "builds"
        ]
      },
      "patch": {
        "operationId": "update_build_builds__build_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Build",
        "tags": [
          "builds"
        ]
      }
    },
    "/builds/{build_id}/page": {
      "put": {
        "operationId": "attach_page_builds__build_id__page_put",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attach Page",
        "tags": [
          "builds"
        ]
      }
    },
...
  }
}